// src/model/response/IResCommercialCredit.ts
export interface IResCommercialCredit {
  id: string;
  name: string;
  date: string;
  plafond: number;
  number: any;
  credit_type: string;
  created_by: string;
  created_date: number;
  type: string;
  request_number: string;
  pk_date: string;
  business_type: string;
  guarantee: string;
  binding_type: string;
  notary: string;
  note: string;
}